🔥 Fix bugs and typo
